+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/CertificationStatusEvent.java
@@ -5,11 +5,11 @@ import java.io.Serializable;
 public class CertificationStatusEvent implements Serializable {
 	private static final long serialVersionUID = -2498656549844148886L;
 	private Long id;
-	private String eventDate;
+	private Long eventDate;
 	private Long certificationStatusId;
 	private String certificationStatusName;
 	private Long lastModifiedUser;
-	private String lastModifiedDate;
+	private Long lastModifiedDate;
 	
 	public CertificationStatusEvent(){}
 	
@@ -19,10 +19,10 @@ public class CertificationStatusEvent implements Serializable {
 	public void setId(Long id) {
 		this.id = id;
 	}
-	public String getEventDate() {
+	public Long getEventDate() {
 		return eventDate;
 	}
-	public void setEventDate(String eventDate) {
+	public void setEventDate(Long eventDate) {
 		this.eventDate = eventDate;
 	}
 	public Long getLastModifiedUser() {
@@ -31,10 +31,10 @@ public class CertificationStatusEvent implements Serializable {
 	public void setLastModifiedUser(Long lastModifiedUser) {
 		this.lastModifiedUser = lastModifiedUser;
 	}
-	public String getLastModifiedDate() {
+	public Long getLastModifiedDate() {
 		return lastModifiedDate;
 	}
-	public void setLastModifiedDate(String lastModifiedDate) {
+	public void setLastModifiedDate(Long lastModifiedDate) {
 		this.lastModifiedDate = lastModifiedDate;
 	}